Working app state
The client is now functional. It connects to the central server to log in. For now all it can do while logged in is sending requests to the central endpoints.

The method of storing keys has been altered, as using TPM is now not viable. Tokens will be stored in local storage, encrypted, through the different OS providers (like DPAPI or Keychain).

Documentation has been added for ease of development later on.